evita_functional_tests/src/test/java/io/evitadb/api/query/order/RandomTest.java
/*
 *
 *                         _ _        ____  ____
 *               _____   _(_) |_ __ _|  _ \| __ )
 *              / _ \ \ / / | __/ _` | | | |  _ \
 *             |  __/\ V /| | || (_| | |_| | |_) |
 *              \___| \_/ |_|\__\__,_|____/|____/
 *
 *   Copyright (c) 2023-2024
 *
 *   Licensed under the Business Source License, Version 1.1 (the "License");
 *   you may not use this file except in compliance with the License.
 *   You may obtain a copy of the License at
 *
 *   https://github.com/FgForrest/evitaDB/blob/master/LICENSE
 *
 *   Unless required by applicable law or agreed to in writing, software
 *   distributed under the License is distributed on an "AS IS" BASIS,
 *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 *   See the License for the specific language governing permissions and
 *   limitations under the License.
 */

package io.evitadb.api.query.order;

import org.junit.jupiter.api.Test;

import static io.evitadb.api.query.QueryConstraints.random;
import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.junit.jupiter.api.Assertions.assertNotNull;
import static org.junit.jupiter.api.Assertions.assertNotSame;
import static org.junit.jupiter.api.Assertions.assertTrue;

/**
 * This tests verifies basic properties of {@link Random} query.
 *
 * @author Jan Novotný (novotny@fg.cz), FG Forrest a.s. (c) 2021
 */
class RandomTest {

	@Test
	void shouldCreateViaFactoryClassWorkAsExpected() {
		final Random random = random();
		assertNotNull(random);
	}

	@Test
	void shouldRecognizeApplicability() {
		assertTrue(random().isApplicable());
	}

	@Test
	void shouldToStringReturnExpectedFormat() {
		final Random random = random();
		assertEquals("random()", random.toString());
	}

	@Test
	void shouldConformToEqualsAndHashContract() {
		assertNotSame(random(), random());
		assertEquals(random(), random());
		assertEquals(random().hashCode(), random().hashCode());
	}

}
